example-python-container/requirements.txt
google-cloud-storage==3.2.0
google-cloud-vision==3.10.2
boto3==1.39.15